+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Journal des modifications
 - Améliorations des doc et scripts d'install.
 - Améliorations de la page d'initialisation de compte utilisateur local.
 - Formulaire de création d'un individu à partir d'un utilisateur : initialisation manquante du nom patronymique.
+- Module StepStar : mise à jour du script bash d'envoi quotidien des thèses vers STEP/STAR.
 - [FIX] Authentification LDAP/CAS : la recherche de l'utilisateur en bdd à partir du username n'utilisait pas le paramètre de config 'ldap_username'.
 - [FIX] Authentification LDAP/CAS : les données d'identité de type \UnicaenApp\Entity\Ldap\People n'étaient pas bien traitées.
 - [FIX] Utilisation du bon code structure pour les établissements (source_code).